test(core): DRY settings env reset and ensure hidden store dir
- Add shared helpers that includes reset_settings_env, teardown_settings_singleton
- Use helpers in CLI & config conftests (autouse)
- Move _ensure_hidden_store_dir to top level conftest
- Add __init__.py for test_config package
