+++ tests/test_core/conftest.py
@@ -13,6 +13,15 @@ from pathlib import Path
 from typer.testing import CliRunner
 
 
+@pytest.fixture(autouse=True)
+def _ensure_hidden_store_dir(tmp_path: Path):
+    d = tmp_path / ".deepeval"
+    d.mkdir(exist_ok=True)
+    # some code expects the file to be there after a run,
+    # but at minimum the directory must exist to avoid FileNotFoundError
+    yield
+
+
 # Silence telemetry for all tests so we don't have to deal with the noise
 @pytest.fixture(autouse=True)
 def _telemetry_opt_out(monkeypatch):
@@ -38,26 +47,3 @@ def hidden_store_dir(tmp_path: Path) -> Path:
     d = tmp_path / ".deepeval"
     d.mkdir(parents=True, exist_ok=True)
     return d
-
-
-@pytest.fixture(autouse=True)
-def _fresh_settings_env(monkeypatch):
-    # Settings is a singleton, so we need to do some cleanup between tests
-    # Reset the singleton so each test gets a fresh Settings instance
-    import deepeval.config.settings as settings_mod
-
-    settings_mod._settings_singleton = None
-
-    # drop any env vars that map to Settings fields to avoid cross test contamination
-    from deepeval.config.settings import Settings
-
-    for k in Settings.model_fields.keys():
-        monkeypatch.delenv(k, raising=False)
-
-    # also ensure no implicit default save path is carried over
-    monkeypatch.delenv("DEEPEVAL_DEFAULT_SAVE", raising=False)
-
-    yield
-
-    # clean after the test too
-    settings_mod._settings_singleton = None
